requirements.txt
streamlit
tensorflow
numpy
pandas
matplotlib
scikit-learn
wfdb
scipy
neurokit2
pdf2image
pytesseract
opencv-python
Pillow